+++ transformer/views.py
@@ -4,7 +4,7 @@ from rest_framework.views import APIView
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.response import Response
 
-from .data_fetchers import ArchivesSpaceDataFetcher, WikidataDataFetcher, WikipediaDataFetcher
+from .fetchers import ArchivesSpaceDataFetcher, WikidataDataFetcher, WikipediaDataFetcher
 from .models import Collection, Object, Agent, Term, TransformRun, FetchRun
 from .serializers import *
 from .transformers import ArchivesSpaceDataTransformer, CartographerDataTransformer, WikidataDataTransformer, WikipediaDataTransformer
